+++ src/main/java/survivalblock/atmosphere/atmospheric_api/not_mixin/datagen/RegistryEntryLookupContainer.java
@@ -14,6 +14,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import survivalblock.atmosphere.atmospheric_api.not_mixin.funny.IsThisEvenNecessary;
 
+@Deprecated(since = "3.3.0")
 @IsThisEvenNecessary(IsThisEvenNecessary.Levels.PROBABLY_NOT)
 @SuppressWarnings("unused")
 public class RegistryEntryLookupContainer {